metsenant/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
lxml==5.1.0
packaging==23.2
PyJWT==2.8.0
pytils==0.4.1
pytz==2023.4
PyYAML==6.0.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tools==0.1.9
uritemplate==4.1.1
